requirement.txt
streamlit>=1.30.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0
numpy>=1.25.0
